splitted auditd test file
