+++ tests/integration/security/compliance/test_disastig_00016.py
@@ -1,0 +1,21 @@
+import pytest
+
+"""
+Ref: SRG-OS-000038-GPOS-00016
+
+Verify the operating system produces audit records containing information to
+establish when (date and time) the events occurred.
+"""
+
+
+@pytest.mark.feature("not container and not lima")
+@pytest.mark.booted(reason="audit event validation requires audit subsystem")
+@pytest.mark.root(reason="required to read audit logs")
+def test_audit_event_contains_timestamp(shell):
+    result = shell(
+        cmd="ausearch -ts recent",
+        capture_output=True,
+    )
+    assert (
+        "time->" in result.stdout or "audit(" in result.stdout
+    ), "stigcompliance: audit records do not contain timestamp information"
