+++ tests/integration/security/compliance/test_disastig_00027.py
@@ -1,0 +1,41 @@
+import pytest
+
+"""
+Ref: SRG-OS-000057-GPOS-00027
+
+Verify the operating system protects audit information from unauthorized read
+access.
+"""
+
+AUDIT_LOG_DIR = "/var/log/audit"
+AUDIT_LOG_FILE = "/var/log/audit/audit.log"
+
+
+@pytest.mark.feature("stig")
+@pytest.mark.booted(reason="requires audit subsystem")
+@pytest.mark.root(reason="required to inspect audit log permissions")
+def test_audit_log_directory_permissions_restricted(file):
+    if file.exists(AUDIT_LOG_DIR):
+        assert file.has_permissions(
+            AUDIT_LOG_DIR, "rwx------"
+        ), f"stigcompliance: audit log directory {AUDIT_LOG_DIR} permissions are not restricted"
+
+
+@pytest.mark.feature("stig")
+@pytest.mark.booted(reason="requires audit subsystem")
+@pytest.mark.root(reason="required to inspect audit log permissions")
+def test_audit_log_file_permissions_restricted(file):
+    if file.exists(AUDIT_LOG_FILE):
+        assert file.has_permissions(
+            AUDIT_LOG_FILE, "rw-------"
+        ), f"stigcompliance: audit log file {AUDIT_LOG_FILE} permissions are not restricted"
+
+
+@pytest.mark.feature("stig")
+@pytest.mark.booted(reason="requires audit subsystem")
+@pytest.mark.root(reason="required to inspect audit log ownership")
+def test_audit_log_owned_by_root(file):
+    if file.exists(AUDIT_LOG_FILE):
+        assert file.is_owned_by_user(
+            AUDIT_LOG_FILE, "root"
+        ), f"stigcompliance: audit log file {AUDIT_LOG_FILE} is not owned by root"
