+++ tests/integration/security/compliance/test_disastig_00028.py
@@ -1,0 +1,12 @@
+import pytest
+
+"""
+Ref: SRG-OS-000058-GPOS-00028   
+
+Verify the operating system protects audit information from unauthorized
+modification.
+"""
+
+
+def test_disastig_00028():
+    pytest.skip(reason="covered by test_disastig_00027.py")
