+++ tests/integration/security/compliance/test_disastig_00191.py
@@ -1,0 +1,20 @@
+import pytest
+
+"""
+Ref: SRG-OS-000432-GPOS-00191
+
+Verify the operating system behaves in a predictable and documented manner that
+reflects organizational and system objectives when invalid inputs are received.
+"""
+
+
+@pytest.mark.feature("not container and not lima")
+@pytest.mark.booted(reason="requires audit subsystem running")
+@pytest.mark.root(reason="required to read audit logs")
+def test_invalid_input_handling_is_audited(shell):
+    result = shell("ausearch -ts recent", capture_output=True)
+    stdout = result.stdout
+
+    assert (
+        "success=no" in stdout or "res=failed" in stdout or "invalid" in stdout.lower()
+    ), "stigcompliance: audit records do not contain evidence of handling invalid inputs"
